+++ topic/topicwriter/topicwriter.go
@@ -66,6 +66,8 @@ func (w *Writer) WaitInitInfo(ctx context.Context) (info PublicInitialInfo, err 
 	return publicInfo, nil
 }
 
+// Close will flush rested messages from buffer and close the writer.
+// You can't write new messages after call Close
 func (w *Writer) Close(ctx context.Context) error {
 	return w.inner.Close(ctx)
 }